Add participants-only visibility and readonly flags for event conversations
Conversation settings: participantsOnlyVisible and participantsReadonly.
Organizers (including org members) bypass both restrictions. Non-admins
cannot see restricted chats or write to readonly chats.

+++ EList.Services.Impl/ConversationService.cs
@@ -20,15 +20,22 @@ namespace EList.Services.Impl
 
         private readonly ICorrelationIdProvider _correlationIdProvider;
         private readonly IConversationRepository _conversationsRepository;
+        private readonly IEventOrganizatorsRepository _eventOrganizatorsRepository;
+        private readonly IParticipationsRepository _participationsRepository;
         private readonly IAccountDataHolder _accountDataHolder;
         private readonly INotificationsService _notificationsService;
+
         public ConversationService(ICorrelationIdProvider correlationIdProvider,
             IConversationRepository conversationsRepository,
+            IEventOrganizatorsRepository eventOrganizatorsRepository,
+            IParticipationsRepository participationsRepository,
             IAccountDataHolder accountDataHolder,
             INotificationsService notificationsService)
         {
             _correlationIdProvider = correlationIdProvider ?? throw new ArgumentNullException(nameof(correlationIdProvider));
             _conversationsRepository = conversationsRepository ?? throw new ArgumentNullException(nameof(conversationsRepository));
+            _eventOrganizatorsRepository = eventOrganizatorsRepository ?? throw new ArgumentNullException(nameof(eventOrganizatorsRepository));
+            _participationsRepository = participationsRepository ?? throw new ArgumentNullException(nameof(participationsRepository));
             _notificationsService = notificationsService ?? throw new ArgumentNullException(nameof(notificationsService));
             _accountDataHolder = accountDataHolder;
         }
@@ -41,7 +48,13 @@ namespace EList.Services.Impl
 
             logger.Debug(correlationId, null, methodName, $"Method started", null);
 
-            //TODO: Добавить проверку что пользователь может создать диалог в рамках текущего события
+            if (conversation.EventId != null)
+            {
+                var isAdmin = await IsEventAdminAsync(conversation.EventId.Value);
+                if (!isAdmin)
+                    return CommandResult<Guid>.Fail(ErrorCode.AccessError, "Создавать диалоги мероприятия может только организатор");
+            }
+
             var result = await _conversationsRepository.CreateConversationAsync(conversation);
 
             logger.Debug(correlationId, null, methodName, $"Method finished", null, execTime.Elapsed);
@@ -56,14 +69,19 @@ namespace EList.Services.Impl
 
             logger.Debug(correlationId, null, methodName, $"Method started", null);
 
-            //TODO: Добавить проверку что пользователь может писать в указанном чате
+            var conversation = await _conversationsRepository.GetConversationAsync(message.ConversationId);
+            if (conversation == null)
+                return CommandResult<Guid>.Fail(ErrorCode.IsNullOrEmpty, "Диалог не найден");
+
+            var writeAccess = await EnsureCanWriteAsync(conversation);
+            if (writeAccess != null)
+                return CommandResult<Guid>.Fail(writeAccess.ErrorCode, writeAccess.Message);
+
+            message.AccountId ??= _accountDataHolder.AccountId;
             var result = await _conversationsRepository.CreateMessageAsync(message);
 
             if (message.ReplyTo != null)
-            {
-                var conversation = await _conversationsRepository.GetConversationAsync(message.ConversationId);
-                await _notificationsService.NotifyCommentRepliedsync(conversation?.EventId, message.ReplyTo.Value, result);
-            }
+                await _notificationsService.NotifyCommentRepliedsync(conversation.EventId, message.ReplyTo.Value, result);
 
             logger.Debug(correlationId, null, methodName, $"Method finished", null, execTime.Elapsed);
             return new CommandResult<Guid>(result);
@@ -77,7 +95,17 @@ namespace EList.Services.Impl
 
             logger.Debug(correlationId, null, methodName, $"Method started", null);
 
-            //TODO: Добавить проверку что пользователь может удалить диалог
+            var conversation = await _conversationsRepository.GetConversationAsync(conversationId);
+            if (conversation == null)
+                return CommandResult.Fail(ErrorCode.IsNullOrEmpty, "Диалог не найден");
+
+            if (conversation.EventId != null)
+            {
+                var isAdmin = await IsEventAdminAsync(conversation.EventId.Value);
+                if (!isAdmin)
+                    return CommandResult.Fail(ErrorCode.AccessError, "Удалять диалоги мероприятия может только организатор");
+            }
+
             await _conversationsRepository.DeleteConversationAsync(conversationId);
 
             logger.Debug(correlationId, null, methodName, $"Method finished", null, execTime.Elapsed);
@@ -93,13 +121,20 @@ namespace EList.Services.Impl
             logger.Debug(correlationId, null, methodName, $"Method started", null);
 
             var existingMessage = await _conversationsRepository.GetMessageAsync(messageId);
-
-            if (existingMessage.AccountId != _accountDataHolder.AccountId)
+            if (existingMessage == null)
                 return CommandResult.Fail(ErrorCode.MessageNotFound, "Сообщение не найдено");
+
+            var conversation = await _conversationsRepository.GetConversationAsync(existingMessage.ConversationId);
+            if (conversation != null)
+            {
+                var writeAccess = await EnsureCanWriteAsync(conversation);
+                if (writeAccess != null)
+                    return writeAccess;
+            }
 
             if (existingMessage.AccountId != _accountDataHolder.AccountId)
                 return CommandResult.Fail(ErrorCode.AccessError, "Нельзя удалять чужие сообщения");
-            
+
             if (existingMessage.Replied)
                 return CommandResult.Fail(ErrorCode.MessageReplied, "Нельзя удалить сообщение на которое уже ответили");
 
@@ -117,10 +152,19 @@ namespace EList.Services.Impl
 
             logger.Debug(correlationId, null, methodName, $"Method started", null);
 
+            if (_accountDataHolder.AccountId == null)
+                return CommandResult<List<Conversation>>.Fail(ErrorCode.AccessError, "Необходимо авторизоваться");
+
             var result = await _conversationsRepository.GetAccountConversationsAsync(_accountDataHolder.AccountId.Value, personalOnly);
+            var visible = new List<Conversation>();
+            foreach (var conversation in result)
+            {
+                if (await CanViewConversationAsync(conversation))
+                    visible.Add(conversation);
+            }
 
             logger.Debug(correlationId, null, methodName, $"Method finished", null, execTime.Elapsed);
-            return new CommandResult<List<Conversation>>(result);
+            return new CommandResult<List<Conversation>>(visible);
         }
 
         public async Task<CommandResult<Conversation?>> GetConversationAsync(Guid conversationId)
@@ -132,6 +176,11 @@ namespace EList.Services.Impl
             logger.Debug(correlationId, null, methodName, $"Method started", null);
 
             var result = await _conversationsRepository.GetConversationAsync(conversationId);
+            if (result == null)
+                return new CommandResult<Conversation?>(null);
+
+            if (!await CanViewConversationAsync(result))
+                return CommandResult<Conversation?>.Fail(ErrorCode.AccessError, "Диалог доступен только участникам мероприятия");
 
             logger.Debug(correlationId, null, methodName, $"Method finished", null, execTime.Elapsed);
             return new CommandResult<Conversation?>(result);
@@ -145,7 +194,13 @@ namespace EList.Services.Impl
 
             logger.Debug(correlationId, null, methodName, $"Method started", null);
 
-            //TODO: Реализовать проверку что пользователь может просматривать сообщения этого чата
+            var conversation = await _conversationsRepository.GetConversationAsync(conversationId);
+            if (conversation == null)
+                return CommandResult<PagedList<Message>>.Fail(ErrorCode.IsNullOrEmpty, "Диалог не найден");
+
+            if (!await CanViewConversationAsync(conversation))
+                return CommandResult<PagedList<Message>>.Fail(ErrorCode.AccessError, "Диалог доступен только участникам мероприятия");
+
             var result = await _conversationsRepository.GetConversationMessagesAsync(conversationId, pageIndex, pageSize);
 
             logger.Debug(correlationId, null, methodName, $"Method finished", null, execTime.Elapsed);
@@ -160,11 +215,16 @@ namespace EList.Services.Impl
 
             logger.Debug(correlationId, null, methodName, $"Method started", null);
 
-            //TODO: Реализовать проверку что пользователь может видеть чаты события (белые, черные списки, закрытое и т.д.)
             var result = await _conversationsRepository.GetEventConversations(eventId);
+            var isAdmin = await IsEventAdminAsync(eventId);
+            var isParticipant = isAdmin || await IsEventParticipantAsync(eventId);
+
+            var visible = result
+                .Where(c => isAdmin || !c.ParticipantsOnlyVisible || isParticipant)
+                .ToList();
 
             logger.Debug(correlationId, null, methodName, $"Method finished", null, execTime.Elapsed);
-            return new CommandResult<List<Conversation>>(result);
+            return new CommandResult<List<Conversation>>(visible);
         }
 
         public async Task<CommandResult<PagedList<Message>>> GetMessageRepliesAsync(Guid messageId, int? pageIndex, int? pageSize)
@@ -175,7 +235,14 @@ namespace EList.Services.Impl
 
             logger.Debug(correlationId, null, methodName, $"Method started", null);
 
-            //TODO: Реализовать проверку что пользователь может видеть эти сообщения (белые, черные списки, закрытое и т.д.)
+            var message = await _conversationsRepository.GetMessageAsync(messageId);
+            if (message == null)
+                return CommandResult<PagedList<Message>>.Fail(ErrorCode.MessageNotFound, "Сообщение не найдено");
+
+            var conversation = await _conversationsRepository.GetConversationAsync(message.ConversationId);
+            if (conversation != null && !await CanViewConversationAsync(conversation))
+                return CommandResult<PagedList<Message>>.Fail(ErrorCode.AccessError, "Диалог доступен только участникам мероприятия");
+
             var result = await _conversationsRepository.GetMessageRepliesAsync(messageId, pageIndex, pageSize);
 
             logger.Debug(correlationId, null, methodName, $"Method finished", null, execTime.Elapsed);
@@ -190,7 +257,21 @@ namespace EList.Services.Impl
 
             logger.Debug(correlationId, null, methodName, $"Method started", null);
 
-            //TODO: Реализовать проверку что пользователь редактировать диалог
+            if (conversation.Id == null)
+                return CommandResult.Fail(ErrorCode.IsNullOrEmpty, "Не указан идентификатор диалога");
+
+            var existing = await _conversationsRepository.GetConversationAsync(conversation.Id.Value);
+            if (existing == null)
+                return CommandResult.Fail(ErrorCode.IsNullOrEmpty, "Диалог не найден");
+
+            var eventId = conversation.EventId ?? existing.EventId;
+            if (eventId != null)
+            {
+                var isAdmin = await IsEventAdminAsync(eventId.Value);
+                if (!isAdmin)
+                    return CommandResult.Fail(ErrorCode.AccessError, "Редактировать диалоги мероприятия может только организатор");
+            }
+
             await _conversationsRepository.UpdateConversationAsync(conversation);
 
             logger.Debug(correlationId, null, methodName, $"Method finished", null, execTime.Elapsed);
@@ -212,6 +293,14 @@ namespace EList.Services.Impl
             if (existingMessage == null)
                 return CommandResult.Fail(ErrorCode.MessageNotFound, $"Сообщение с id='{message.Id}' не найдено");
 
+            var conversation = await _conversationsRepository.GetConversationAsync(existingMessage.ConversationId);
+            if (conversation != null)
+            {
+                var writeAccess = await EnsureCanWriteAsync(conversation);
+                if (writeAccess != null)
+                    return writeAccess;
+            }
+
             if (existingMessage.AccountId != _accountDataHolder.AccountId)
                 return CommandResult.Fail(ErrorCode.AccessError, $"Нельзя редактировать сообщения другого пользователя");
 
@@ -219,6 +308,55 @@ namespace EList.Services.Impl
 
             logger.Debug(correlationId, null, methodName, $"Method finished", null, execTime.Elapsed);
             return CommandResult.OK;
+        }
+
+        private async Task<bool> IsEventAdminAsync(Guid eventId)
+        {
+            if (_accountDataHolder.AccountId == null)
+                return false;
+
+            return await _eventOrganizatorsRepository.IsAccountEventOrganizatorAsync(eventId, _accountDataHolder.AccountId.Value);
+        }
+
+        private async Task<bool> IsEventParticipantAsync(Guid eventId)
+        {
+            if (_accountDataHolder.AccountId == null)
+                return false;
+
+            return await _participationsRepository.IsUserParticipatedAsync(_accountDataHolder.AccountId.Value, eventId);
+        }
+
+        private async Task<bool> CanViewConversationAsync(Conversation conversation)
+        {
+            if (conversation.EventId == null || !conversation.ParticipantsOnlyVisible)
+                return true;
+
+            if (await IsEventAdminAsync(conversation.EventId.Value))
+                return true;
+
+            return await IsEventParticipantAsync(conversation.EventId.Value);
+        }
+
+        private async Task<CommandResult?> EnsureCanWriteAsync(Conversation conversation)
+        {
+            if (conversation.EventId == null)
+                return null;
+
+            var isAdmin = await IsEventAdminAsync(conversation.EventId.Value);
+            if (isAdmin)
+                return null;
+
+            if (conversation.ParticipantsOnlyVisible)
+            {
+                var isParticipant = await IsEventParticipantAsync(conversation.EventId.Value);
+                if (!isParticipant)
+                    return CommandResult.Fail(ErrorCode.AccessError, "Диалог доступен только участникам мероприятия");
+            }
+
+            if (conversation.ParticipantsReadonly)
+                return CommandResult.Fail(ErrorCode.AccessError, "Участники могут только читать сообщения в этом диалоге");
+
+            return null;
         }
     }
 }
